# [GraphBolt] Update `ItemSampler`