--- a/python/dgl/graphbolt/itemset.py
+++ b/python/dgl/graphbolt/itemset.py
@@ -362,7 +362,7 @@ class ItemSetDict:
                     break
             return data
         elif isinstance(index, Iterable):
-            index = torch.tensor(index).squeeze()
+            index = torch.tensor(index)
             assert torch.all((index >= 0) & (index < self._length))
             key_indices = (
                 torch.searchsorted(self._offsets, index, right=True) - 1