## Description
<!-- Brief description. Refer to the related issues if existed.
It'll be great if relevant reviewers can be assigned as well.-->

1. Update `ItemSampler` to support correct stochastic sharding across distributed groups.
2. Modify the logic of `ItemSet.__getitem__()` when index is an iterable of int.

Benchmark on `examples/multigpu/graphbolt/node_classification.py`.

Old:
```
python /home/ubuntu/dgl/examples/multigpu/graphbolt/node_classification.py --gpu 0,1,2,3
Training with 4 gpus.
The dataset is already preprocessed.
Training...
48it [00:02, 16.06it/s]
Validating...
10it [00:00, 21.67it/s]
Epoch 00000 | Average Loss 2.3267 | Accuracy 0.7917 | Time 3.5637
48it [00:02, 21.37it/s]
Validating...
10it [00:00, 24.19it/s]
Epoch 00001 | Average Loss 0.9559 | Accuracy 0.8437 | Time 2.7528
48it [00:02, 21.33it/s]
Validating...
10it [00:00, 24.10it/s]
Epoch 00002 | Average Loss 0.7238 | Accuracy 0.8602 | Time 2.7597
48it [00:02, 21.33it/s]
Validating...
10it [00:00, 24.51it/s]
Epoch 00003 | Average Loss 0.6163 | Accuracy 0.8706 | Time 2.7502
48it [00:02, 21.45it/s]
Validating...
10it [00:00, 24.45it/s]
Epoch 00004 | Average Loss 0.5578 | Accuracy 0.8762 | Time 2.7404
48it [00:02, 20.19it/s]
Validating...
10it [00:00, 24.57it/s]
Epoch 00005 | Average Loss 0.5176 | Accuracy 0.8819 | Time 2.8776
48it [00:02, 21.50it/s]
Validating...
10it [00:00, 24.13it/s]
Epoch 00006 | Average Loss 0.4883 | Accuracy 0.8855 | Time 2.7396
48it [00:02, 21.42it/s]
Validating...
10it [00:00, 24.41it/s]
Epoch 00007 | Average Loss 0.4667 | Accuracy 0.8881 | Time 2.7437
48it [00:02, 21.31it/s]
Validating...
10it [00:00, 24.19it/s]
Epoch 00008 | Average Loss 0.4477 | Accuracy 0.8889 | Time 2.7596
48it [00:02, 21.46it/s]
Validating...
10it [00:00, 24.29it/s]
Epoch 00009 | Average Loss 0.4343 | Accuracy 0.8920 | Time 2.7416
Testing...
541it [00:19, 27.95it/s]
Test Accuracy 0.7348
```

New:
```
python /home/ubuntu/dgl/examples/multigpu/graphbolt/node_classification.py --gpu 0,1,2,3
Training with 4 gpus.
The dataset is already preprocessed.
Training...
48it [00:02, 16.78it/s]
Validating...
10it [00:00, 21.44it/s]
Epoch 00000 | Average Loss 2.4174 | Accuracy 0.7767 | Time 3.4403
48it [00:02, 21.43it/s]
Validating...
10it [00:00, 24.34it/s]
Epoch 00001 | Average Loss 0.9978 | Accuracy 0.8386 | Time 2.7454
48it [00:02, 21.31it/s]
Validating...
10it [00:00, 24.78it/s]
Epoch 00002 | Average Loss 0.7490 | Accuracy 0.8564 | Time 2.7504
48it [00:02, 21.46it/s]
Validating...
10it [00:00, 24.60it/s]
Epoch 00003 | Average Loss 0.6361 | Accuracy 0.8688 | Time 2.7368
48it [00:02, 21.41it/s]
Validating...
10it [00:00, 24.71it/s]
Epoch 00004 | Average Loss 0.5715 | Accuracy 0.8754 | Time 2.7415
48it [00:02, 20.55it/s]
Validating...
10it [00:00, 24.19it/s]
Epoch 00005 | Average Loss 0.5289 | Accuracy 0.8806 | Time 2.8437
48it [00:02, 21.36it/s]
Validating...
10it [00:00, 24.73it/s]
Epoch 00006 | Average Loss 0.5007 | Accuracy 0.8829 | Time 2.7457
48it [00:02, 21.33it/s]
Validating...
10it [00:00, 24.50it/s]
Epoch 00007 | Average Loss 0.4760 | Accuracy 0.8861 | Time 2.7551
48it [00:02, 21.36it/s]
Validating...
10it [00:00, 24.57it/s]
Epoch 00008 | Average Loss 0.4555 | Accuracy 0.8883 | Time 2.7549
48it [00:02, 21.41it/s]
Validating...
10it [00:00, 24.54it/s]
Epoch 00009 | Average Loss 0.4404 | Accuracy 0.8897 | Time 2.7446
Testing...
541it [00:19, 27.94it/s]
Test Accuracy 0.7326
```

## Checklist
Please feel free to remove inapplicable items for your PR.
- [ ] The PR title starts with [$CATEGORY] (such as [NN], [Model], [Doc], [Feature]])
- [ ] I've leverage the [tools](https://docs.google.com/document/d/1iHyj7zlmygKSk5gBPsqIqL5ASPzJSPREaNT_QdsiYA4/edit) to beautify the python and c++ code.
- [ ] The PR is complete and small, read the [Google eng practice (CL equals to PR)](https://google.github.io/eng-practices/review/developer/small-cls.html) to understand more about small PR. In DGL, we consider PRs with less than 200 lines of core code change are small (example, test and documentation could be exempted).
- [ ] All changes have test coverage
- [ ] Code is well-documented
- [ ] To the best of my knowledge, examples are either not affected by this change, or have been fixed to be compatible with this change
- [ ] Related issue is referred in this PR
- [ ] If the PR is for a new model/paper, I've updated the example index [here](../examples/README.md).

## Changes
<!-- You could use following template
- [ ] Feature1, tests, (and when applicable, API doc)
- [ ] Feature2, tests, (and when applicable, API doc)
-->