@Rhett-Ying 2024-05-21
@Skeleton003 According to the doc, `examples/sampling/graphbolt/link_prediction.py` runs a bit slower. Please keep an eye on the regression results.